+++ package.js
@@ -1,33 +1,35 @@
-/* global Package, Npm */
+/* eslint-env meteor */
 
 Package.describe({
   summary: 'Authorization package for Meteor',
-  version: '4.0.0',
+  version: '4.0.0-alpha.3',
   git: 'https://github.com/Meteor-Community-Packages/meteor-roles.git',
   name: 'alanning:roles'
 })
 
 Package.onUse(function (api) {
-  api.versionsFrom(['2.8.0'])
+  api.versionsFrom(['2.8.1', '3.0-rc.0'])
 
-  var both = ['client', 'server']
+  const both = ['client', 'server']
 
   api.use([
     'ecmascript',
     'accounts-base',
     'tracker',
     'mongo',
-    'check'
+    'check',
+    'ddp'
   ], both)
 
-  api.use('zodern:types@1.0.9')
+  api.use('zodern:types@1.0.13')
 
-  api.use(['blaze@2.7.1'], 'client', { weak: true })
+  api.use(['blaze@2.9.0'], 'client', { weak: true })
 
   api.export(['Roles', 'RolesCollection', 'RoleAssignmentCollection'])
 
-  api.addFiles('roles/roles_common.js', both)
-  api.mainModule('roles/roles_server.js', 'server')
+  api.addFiles('roles/roles_client.js', 'client')
+  api.addFiles('roles/roles_common_async.js', both)
+  api.addFiles('roles/roles_server.js', 'server')
   api.addFiles([
     'roles/client/debug.js',
     'roles/client/uiHelpers.js',
@@ -38,18 +40,14 @@ Package.onUse(function (api) {
 Package.onTest(function (api) {
   // Add code coverage
   api.use([
-    'lmieulet:meteor-packages-coverage@0.2.0',
-    'lmieulet:meteor-coverage@3.2.0',
-    'meteortesting:mocha'
+    'lmieulet:meteor-legacy-coverage',
+    'lmieulet:meteor-coverage@4.1.0',
+    'meteortesting:mocha@2.1.0'
   ])
 
-  Npm.depends({
-    'chai': '4.2.0'
-  })
+  api.versionsFrom('2.8.1')
 
-  api.versionsFrom('2.3')
-
-  var both = ['client', 'server']
+  const both = ['client', 'server']
 
   // `accounts-password` is included so `Meteor.users` exists
 
@@ -59,6 +57,7 @@ Package.onTest(function (api) {
     'mongo'
   ], both)
 
+  api.addFiles('roles/tests/serverAsync.js', 'server')
   api.addFiles('roles/tests/client.js', 'client')
-  api.addFiles('roles/tests/server.js', 'server')
+  api.addFiles('roles/tests/clientAsync.js', 'client')
 })