+++ roles/roles_common_async.js
@@ -1,4 +1,6 @@
-/* global Meteor, Roles, Mongo */
+/* global Roles */
+import { Meteor } from 'meteor/meteor'
+import { Mongo } from 'meteor/mongo'
 
 /**
  * Provides functions related to user authorization. Compatible with built-in Meteor accounts packages.
@@ -24,7 +26,6 @@
  *
  * @module Roles
  */
-
 export const RolesCollection = new Mongo.Collection('roles')
 
 if (!Meteor.roles) {
@@ -44,10 +45,22 @@ if (typeof Roles === 'undefined') {
   Roles = {} // eslint-disable-line no-global-assign
 }
 
-var getGroupsForUserDeprecationWarning = false
+let getGroupsForUserDeprecationWarning = false
+
+/**
+ * Helper, resolves async some
+ * @param {*} arr
+ * @param {*} predicate
+ * @returns {Promise<Boolean>}
+ */
+const asyncSome = async (arr, predicate) => {
+  for (const e of arr) {
+    if (await predicate(e)) return true
+  }
+  return false
+}
 
 Object.assign(Roles, {
-
   /**
    * Used as a global group (now scope) name. Not used anymore.
    *
@@ -60,28 +73,46 @@ Object.assign(Roles, {
   /**
    * Create a new role.
    *
-   * @method createRole
+   * @method createRoleAsync
    * @param {String} roleName Name of role.
    * @param {Object} [options] Options:
    *   - `unlessExists`: if `true`, exception will not be thrown in the role already exists
-   * @return {String} ID of the new role or null.
+   * @return {Promise<String>} ID of the new role or null.
    * @static
    */
-  createRole: function (roleName, options) {
+  createRoleAsync: async function (roleName, options) {
     Roles._checkRoleName(roleName)
 
-    options = Object.assign({
-      unlessExists: false
-    }, options)
+    options = Object.assign(
+      {
+        unlessExists: false
+      },
+      options
+    )
 
-    var result = Meteor.roles.upsert({ _id: roleName }, { $setOnInsert: { children: [] } })
+    let insertedId = null
 
-    if (!result.insertedId) {
-      if (options.unlessExists) return null
-      throw new Error('Role \'' + roleName + '\' already exists.')
+    const existingRole = await Meteor.roles.findOneAsync({ _id: roleName })
+
+    if (existingRole) {
+      await Meteor.roles.updateAsync(
+        { _id: roleName },
+        { $setOnInsert: { children: [] } }
+      )
+      return null
+    } else {
+      insertedId = await Meteor.roles.insertAsync({
+        _id: roleName,
+        children: []
+      })
     }
 
-    return result.insertedId
+    if (!insertedId) {
+      if (options.unlessExists) return null
+      throw new Error("Role '" + roleName + "' already exists.")
+    }
+
+    return insertedId
   },
 
   /**
@@ -89,109 +120,137 @@ Object.assign(Roles, {
    *
    * If the role is set for any user, it is automatically unset.
    *
-   * @method deleteRole
+   * @method deleteRoleAsync
    * @param {String} roleName Name of role.
+   * @returns {Promise}
    * @static
    */
-  deleteRole: function (roleName) {
-    var roles
-    var inheritedRoles
+  deleteRoleAsync: async function (roleName) {
+    let roles
+    let inheritedRoles
 
     Roles._checkRoleName(roleName)
 
     // Remove all assignments
-    Meteor.roleAssignment.remove({
+    await Meteor.roleAssignment.removeAsync({
       'role._id': roleName
     })
 
     do {
       // For all roles who have it as a dependency ...
-      roles = Roles._getParentRoleNames(Meteor.roles.findOne({ _id: roleName }))
+      roles = await Roles._getParentRoleNamesAsync(
+        await Meteor.roles.findOneAsync({ _id: roleName })
+      )
 
-      Meteor.roles.find({ _id: { $in: roles } }).fetch().forEach(r => {
-        Meteor.roles.update({
-          _id: r._id
-        }, {
-          $pull: {
-            children: {
-              _id: roleName
+      for (const r of await Meteor.roles
+        .find({ _id: { $in: roles } })
+        .fetchAsync()) {
+        await Meteor.roles.updateAsync(
+          {
+            _id: r._id
+          },
+          {
+            $pull: {
+              children: {
+                _id: roleName
+              }
             }
           }
-        })
+        )
 
-        inheritedRoles = Roles._getInheritedRoleNames(Meteor.roles.findOne({ _id: r._id }))
-        Meteor.roleAssignment.update({
-          'role._id': r._id
-        }, {
-          $set: {
-            inheritedRoles: [r._id, ...inheritedRoles].map(r2 => ({ _id: r2 }))
-          }
-        }, { multi: true })
-      })
+        inheritedRoles = await Roles._getInheritedRoleNamesAsync(
+          await Meteor.roles.findOneAsync({ _id: r._id })
+        )
+        await Meteor.roleAssignment.updateAsync(
+          {
+            'role._id': r._id
+          },
+          {
+            $set: {
+              inheritedRoles: [r._id, ...inheritedRoles].map((r2) => ({
+                _id: r2
+              }))
+            }
+          },
+          { multi: true }
+        )
+      }
     } while (roles.length > 0)
 
     // And finally remove the role itself
-    Meteor.roles.remove({ _id: roleName })
+    await Meteor.roles.removeAsync({ _id: roleName })
   },
 
   /**
    * Rename an existing role.
    *
-   * @method renameRole
+   * @method renameRoleAsync
    * @param {String} oldName Old name of a role.
    * @param {String} newName New name of a role.
+   * @returns {Promise}
    * @static
    */
-  renameRole: function (oldName, newName) {
-    var role
-    var count
+  renameRoleAsync: async function (oldName, newName) {
+    let count
 
     Roles._checkRoleName(oldName)
     Roles._checkRoleName(newName)
 
     if (oldName === newName) return
 
-    role = Meteor.roles.findOne({ _id: oldName })
+    const role = await Meteor.roles.findOneAsync({ _id: oldName })
 
     if (!role) {
-      throw new Error('Role \'' + oldName + '\' does not exist.')
+      throw new Error("Role '" + oldName + "' does not exist.")
     }
 
     role._id = newName
 
-    Meteor.roles.insert(role)
+    await Meteor.roles.insertAsync(role)
 
     do {
-      count = Meteor.roleAssignment.update({
-        'role._id': oldName
-      }, {
-        $set: {
-          'role._id': newName
-        }
-      }, { multi: true })
+      count = await Meteor.roleAssignment.updateAsync(
+        {
+          'role._id': oldName
+        },
+        {
+          $set: {
+            'role._id': newName
+          }
+        },
+        { multi: true }
+      )
     } while (count > 0)
 
     do {
-      count = Meteor.roleAssignment.update({
-        'inheritedRoles._id': oldName
-      }, {
-        $set: {
-          'inheritedRoles.$._id': newName
-        }
-      }, { multi: true })
+      count = await Meteor.roleAssignment.updateAsync(
+        {
+          'inheritedRoles._id': oldName
+        },
+        {
+          $set: {
+            'inheritedRoles.$._id': newName
+          }
+        },
+        { multi: true }
+      )
     } while (count > 0)
 
     do {
-      count = Meteor.roles.update({
-        'children._id': oldName
-      }, {
-        $set: {
-          'children.$._id': newName
-        }
-      }, { multi: true })
+      count = await Meteor.roles.updateAsync(
+        {
+          'children._id': oldName
+        },
+        {
+          $set: {
+            'children.$._id': newName
+          }
+        },
+        { multi: true }
+      )
     } while (count > 0)
 
-    Meteor.roles.remove({ _id: oldName })
+    await Meteor.roles.removeAsync({ _id: oldName })
   },
 
   /**
@@ -200,70 +259,83 @@ Object.assign(Roles, {
    * Previous parents are kept (role can have multiple parents). For users which have the
    * parent role set, new subroles are added automatically.
    *
-   * @method addRolesToParent
+   * @method addRolesToParentAsync
    * @param {Array|String} rolesNames Name(s) of role(s).
    * @param {String} parentName Name of parent role.
+   * @returns {Promise}
    * @static
    */
-  addRolesToParent: function (rolesNames, parentName) {
+  addRolesToParentAsync: async function (rolesNames, parentName) {
     // ensure arrays
     if (!Array.isArray(rolesNames)) rolesNames = [rolesNames]
 
-    rolesNames.forEach(function (roleName) {
-      Roles._addRoleToParent(roleName, parentName)
-    })
+    for (const roleName of rolesNames) {
+      await Roles._addRoleToParentAsync(roleName, parentName)
+    }
   },
 
   /**
-   * @method _addRoleToParent
+   * @method _addRoleToParentAsync
    * @param {String} roleName Name of role.
    * @param {String} parentName Name of parent role.
+   * @returns {Promise}
    * @private
    * @static
    */
-  _addRoleToParent: function (roleName, parentName) {
-    var role
-    var count
-
+  _addRoleToParentAsync: async function (roleName, parentName) {
     Roles._checkRoleName(roleName)
     Roles._checkRoleName(parentName)
 
     // query to get role's children
-    role = Meteor.roles.findOne({ _id: roleName })
+    const role = await Meteor.roles.findOneAsync({ _id: roleName })
 
     if (!role) {
-      throw new Error('Role \'' + roleName + '\' does not exist.')
+      throw new Error("Role '" + roleName + "' does not exist.")
     }
 
     // detect cycles
-    if (Roles._getInheritedRoleNames(role).includes(parentName)) {
-      throw new Error('Roles \'' + roleName + '\' and \'' + parentName + '\' would form a cycle.')
+    if ((await Roles._getInheritedRoleNamesAsync(role)).includes(parentName)) {
+      throw new Error(
+        "Roles '" + roleName + "' and '" + parentName + "' would form a cycle."
+      )
     }
 
-    count = Meteor.roles.update({
-      _id: parentName,
-      'children._id': {
-        $ne: role._id
-      }
-    }, {
-      $push: {
-        children: {
-          _id: role._id
+    const count = await Meteor.roles.updateAsync(
+      {
+        _id: parentName,
+        'children._id': {
+          $ne: role._id
+        }
+      },
+      {
+        $push: {
+          children: {
+            _id: role._id
+          }
         }
       }
-    })
+    )
 
     // if there was no change, parent role might not exist, or role is
-    // already a subrole; in any case we do not have anything more to do
+    // already a sub-role; in any case we do not have anything more to do
     if (!count) return
 
-    Meteor.roleAssignment.update({
-      'inheritedRoles._id': parentName
-    }, {
-      $push: {
-        inheritedRoles: { $each: [role._id, ...Roles._getInheritedRoleNames(role)].map(r => ({ _id: r })) }
-      }
-    }, { multi: true })
+    await Meteor.roleAssignment.updateAsync(
+      {
+        'inheritedRoles._id': parentName
+      },
+      {
+        $push: {
+          inheritedRoles: {
+            $each: [
+              role._id,
+              ...(await Roles._getInheritedRoleNamesAsync(role))
+            ].map((r) => ({ _id: r }))
+          }
+        }
+      },
+      { multi: true }
+    )
   },
 
   /**
@@ -272,67 +344,90 @@ Object.assign(Roles, {
    * Other parents are kept (role can have multiple parents). For users which have the
    * parent role set, removed subrole is removed automatically.
    *
-   * @method removeRolesFromParent
+   * @method removeRolesFromParentAsync
    * @param {Array|String} rolesNames Name(s) of role(s).
    * @param {String} parentName Name of parent role.
+   * @returns {Promise}
    * @static
    */
-  removeRolesFromParent: function (rolesNames, parentName) {
+  removeRolesFromParentAsync: async function (rolesNames, parentName) {
     // ensure arrays
     if (!Array.isArray(rolesNames)) rolesNames = [rolesNames]
 
-    rolesNames.forEach(function (roleName) {
-      Roles._removeRoleFromParent(roleName, parentName)
-    })
+    for (const roleName of rolesNames) {
+      await Roles._removeRoleFromParentAsync(roleName, parentName)
+    }
   },
 
   /**
-   * @method _removeRoleFromParent
+   * @method _removeRoleFromParentAsync
    * @param {String} roleName Name of role.
    * @param {String} parentName Name of parent role.
+   * @returns {Promise}
    * @private
    * @static
    */
-  _removeRoleFromParent: function (roleName, parentName) {
+  _removeRoleFromParentAsync: async function (roleName, parentName) {
     Roles._checkRoleName(roleName)
     Roles._checkRoleName(parentName)
 
     // check for role existence
     // this would not really be needed, but we are trying to match addRolesToParent
-    let role = Meteor.roles.findOne({ _id: roleName }, { fields: { _id: 1 } })
+    const role = await Meteor.roles.findOneAsync(
+      { _id: roleName },
+      { fields: { _id: 1 } }
+    )
 
     if (!role) {
-      throw new Error('Role \'' + roleName + '\' does not exist.')
+      throw new Error("Role '" + roleName + "' does not exist.")
     }
 
-    const count = Meteor.roles.update({
-      _id: parentName
-    }, {
-      $pull: {
-        children: {
-          _id: role._id
+    const count = await Meteor.roles.updateAsync(
+      {
+        _id: parentName
+      },
+      {
+        $pull: {
+          children: {
+            _id: role._id
+          }
         }
       }
-    })
+    )
 
     // if there was no change, parent role might not exist, or role was
     // already not a subrole; in any case we do not have anything more to do
     if (!count) return
 
     // For all roles who have had it as a dependency ...
-    const roles = [...Roles._getParentRoleNames(Meteor.roles.findOne({ _id: parentName })), parentName]
+    const roles = [
+      ...(await Roles._getParentRoleNamesAsync(
+        await Meteor.roles.findOneAsync({ _id: parentName })
+      )),
+      parentName
+    ]
 
-    Meteor.roles.find({ _id: { $in: roles } }).fetch().forEach(r => {
-      const inheritedRoles = Roles._getInheritedRoleNames(Meteor.roles.findOne({ _id: r._id }))
-      Meteor.roleAssignment.update({
-        'role._id': r._id,
-        'inheritedRoles._id': role._id
-      }, {
-        $set: {
-          inheritedRoles: [r._id, ...inheritedRoles].map(r2 => ({ _id: r2 }))
-        }
-      }, { multi: true })
-    })
+    for (const r of await Meteor.roles
+      .find({ _id: { $in: roles } })
+      .fetchAsync()) {
+      const inheritedRoles = await Roles._getInheritedRoleNamesAsync(
+        await Meteor.roles.findOneAsync({ _id: r._id })
+      )
+      await Meteor.roleAssignment.updateAsync(
+        {
+          'role._id': r._id,
+          'inheritedRoles._id': role._id
+        },
+        {
+          $set: {
+            inheritedRoles: [r._id, ...inheritedRoles].map((r2) => ({
+              _id: r2
+            }))
+          }
+        },
+        { multi: true }
+      )
+    }
   },
 
   /**
@@ -341,26 +436,27 @@ Object.assign(Roles, {
    * Adds roles to existing roles for each user.
    *
    * @example
-   *     Roles.addUsersToRoles(userId, 'admin')
-   *     Roles.addUsersToRoles(userId, ['view-secrets'], 'example.com')
-   *     Roles.addUsersToRoles([user1, user2], ['user','editor'])
-   *     Roles.addUsersToRoles([user1, user2], ['glorious-admin', 'perform-action'], 'example.org')
+   *     Roles.addUsersToRolesAsync(userId, 'admin')
+   *     Roles.addUsersToRolesAsync(userId, ['view-secrets'], 'example.com')
+   *     Roles.addUsersToRolesAsync([user1, user2], ['user','editor'])
+   *     Roles.addUsersToRolesAsync([user1, user2], ['glorious-admin', 'perform-action'], 'example.org')
    *
-   * @method addUsersToRoles
+   * @method addUsersToRolesAsync
    * @param {Array|String} users User ID(s) or object(s) with an `_id` field.
    * @param {Array|String} roles Name(s) of roles to add users to. Roles have to exist.
    * @param {Object|String} [options] Options:
    *   - `scope`: name of the scope, or `null` for the global role
    *   - `ifExists`: if `true`, do not throw an exception if the role does not exist
+   * @returns {Promise}
    *
    * Alternatively, it can be a scope name string.
    * @static
    */
-  addUsersToRoles: function (users, roles, options) {
-    var id
+  addUsersToRolesAsync: async function (users, roles, options) {
+    let id
 
-    if (!users) throw new Error('Missing \'users\' param.')
-    if (!roles) throw new Error('Missing \'roles\' param.')
+    if (!users) throw new Error("Missing 'users' param.")
+    if (!roles) throw new Error("Missing 'roles' param.")
 
     options = Roles._normalizeOptions(options)
 
@@ -370,21 +466,24 @@ Object.assign(Roles, {
 
     Roles._checkScopeName(options.scope)
 
-    options = Object.assign({
-      ifExists: false
-    }, options)
+    options = Object.assign(
+      {
+        ifExists: false
+      },
+      options
+    )
 
-    users.forEach(function (user) {
+    for (const user of users) {
       if (typeof user === 'object') {
         id = user._id
       } else {
         id = user
       }
 
-      roles.forEach(function (role) {
-        Roles._addUserToRole(id, role, options)
-      })
-    })
+      for (const role of roles) {
+        await Roles._addUserToRoleAsync(id, role, options)
+      }
+    }
   },
 
   /**
@@ -393,27 +492,28 @@ Object.assign(Roles, {
    * Replaces all existing roles with a new set of roles.
    *
    * @example
-   *     Roles.setUserRoles(userId, 'admin')
-   *     Roles.setUserRoles(userId, ['view-secrets'], 'example.com')
-   *     Roles.setUserRoles([user1, user2], ['user','editor'])
-   *     Roles.setUserRoles([user1, user2], ['glorious-admin', 'perform-action'], 'example.org')
+   *     await Roles.setUserRolesAsync(userId, 'admin')
+   *     await Roles.setUserRolesAsync(userId, ['view-secrets'], 'example.com')
+   *     await Roles.setUserRolesAsync([user1, user2], ['user','editor'])
+   *     await Roles.setUserRolesAsync([user1, user2], ['glorious-admin', 'perform-action'], 'example.org')
    *
-   * @method setUserRoles
+   * @method setUserRolesAsync
    * @param {Array|String} users User ID(s) or object(s) with an `_id` field.
    * @param {Array|String} roles Name(s) of roles to add users to. Roles have to exist.
    * @param {Object|String} [options] Options:
    *   - `scope`: name of the scope, or `null` for the global role
    *   - `anyScope`: if `true`, remove all roles the user has, of any scope, if `false`, only the one in the same scope
    *   - `ifExists`: if `true`, do not throw an exception if the role does not exist
+   * @returns {Promise}
    *
    * Alternatively, it can be a scope name string.
    * @static
    */
-  setUserRoles: function (users, roles, options) {
-    var id
+  setUserRolesAsync: async function (users, roles, options) {
+    let id
 
-    if (!users) throw new Error('Missing \'users\' param.')
-    if (!roles) throw new Error('Missing \'roles\' param.')
+    if (!users) throw new Error("Missing 'users' param.")
+    if (!roles) throw new Error("Missing 'roles' param.")
 
     options = Roles._normalizeOptions(options)
 
@@ -423,12 +523,15 @@ Object.assign(Roles, {
 
     Roles._checkScopeName(options.scope)
 
-    options = Object.assign({
-      ifExists: false,
-      anyScope: false
-    }, options)
+    options = Object.assign(
+      {
+        ifExists: false,
+        anyScope: false
+      },
+      options
+    )
 
-    users.forEach(function (user) {
+    for (const user of users) {
       if (typeof user === 'object') {
         id = user._id
       } else {
@@ -440,28 +543,29 @@ Object.assign(Roles, {
         selector.scope = options.scope
       }
 
-      Meteor.roleAssignment.remove(selector)
+      await Meteor.roleAssignment.removeAsync(selector)
 
       // and then add all
-      roles.forEach(function (role) {
-        Roles._addUserToRole(id, role, options)
-      })
-    })
+      for (const role of roles) {
+        await Roles._addUserToRoleAsync(id, role, options)
+      }
+    }
   },
 
   /**
    * Add one user to one role.
    *
-   * @method _addUserToRole
+   * @method _addUserToRoleAsync
    * @param {String} userId The user ID.
    * @param {String} roleName Name of the role to add the user to. The role have to exist.
    * @param {Object} options Options:
    *   - `scope`: name of the scope, or `null` for the global role
    *   - `ifExists`: if `true`, do not throw an exception if the role does not exist
+   * @returns {Promise}
    * @private
    * @static
    */
-  _addUserToRole: function (userId, roleName, options) {
+  _addUserToRoleAsync: async function (userId, roleName, options) {
     Roles._checkRoleName(roleName)
     Roles._checkScopeName(options.scope)
 
@@ -469,36 +573,77 @@ Object.assign(Roles, {
       return
     }
 
-    const role = Meteor.roles.findOne({ _id: roleName }, { fields: { children: 1 } })
+    const role = await Meteor.roles.findOneAsync(
+      { _id: roleName },
+      { fields: { children: 1 } }
+    )
 
     if (!role) {
       if (options.ifExists) {
         return []
       } else {
-        throw new Error('Role \'' + roleName + '\' does not exist.')
+        throw new Error("Role '" + roleName + "' does not exist.")
       }
     }
 
     // This might create duplicates, because we don't have a unique index, but that's all right. In case there are two, withdrawing the role will effectively kill them both.
-    const res = Meteor.roleAssignment.upsert({
+    // TODO revisit this
+    /* const res = await Meteor.roleAssignment.upsertAsync(
+      {
+        "user._id": userId,
+        "role._id": roleName,
+        scope: options.scope,
+      },
+      {
+        $setOnInsert: {
+          user: { _id: userId },
+          role: { _id: roleName },
+          scope: options.scope,
+        },
+      }
+    ); */
+    const existingAssignment = await Meteor.roleAssignment.findOneAsync({
       'user._id': userId,
       'role._id': roleName,
       scope: options.scope
-    }, {
-      $setOnInsert: {
+    })
+
+    let insertedId
+    let res
+    if (existingAssignment) {
+      await Meteor.roleAssignment.updateAsync(existingAssignment._id, {
+        $set: {
+          user: { _id: userId },
+          role: { _id: roleName },
+          scope: options.scope
+        }
+      })
+
+      res = await Meteor.roleAssignment.findOneAsync(existingAssignment._id)
+    } else {
+      insertedId = await Meteor.roleAssignment.insertAsync({
         user: { _id: userId },
         role: { _id: roleName },
         scope: options.scope
-      }
-    })
-
-    if (res.insertedId) {
-      Meteor.roleAssignment.update({ _id: res.insertedId }, {
-        $set: {
-          inheritedRoles: [roleName, ...Roles._getInheritedRoleNames(role)].map(r => ({ _id: r }))
-        }
       })
     }
+
+    if (insertedId) {
+      await Meteor.roleAssignment.updateAsync(
+        { _id: insertedId },
+        {
+          $set: {
+            inheritedRoles: [
+              roleName,
+              ...(await Roles._getInheritedRoleNamesAsync(role))
+            ].map((r) => ({ _id: r }))
+          }
+        }
+      )
+
+      res = await Meteor.roleAssignment.findOneAsync({ _id: insertedId })
+    }
+    res.insertedId = insertedId // For backward compatibility
 
     return res
   },
@@ -507,27 +652,28 @@ Object.assign(Roles, {
    * Returns an array of role names the given role name is a child of.
    *
    * @example
-   *     Roles._getParentRoleNames({ _id: 'admin', children; [] })
+   *     Roles._getParentRoleNamesAsync({ _id: 'admin', children; [] })
    *
-   * @method _getParentRoleNames
+   * @method _getParentRoleNamesAsync
    * @param {object} role The role object
+   * @returns {Promise}
    * @private
    * @static
    */
-  _getParentRoleNames: function (role) {
-    var parentRoles
-
+  _getParentRoleNamesAsync: async function (role) {
     if (!role) {
       return []
     }
 
-    parentRoles = new Set([role._id])
+    const parentRoles = new Set([role._id])
 
-    parentRoles.forEach(roleName => {
-      Meteor.roles.find({ 'children._id': roleName }).fetch().forEach(parentRole => {
+    for (const roleName of parentRoles) {
+      for (const parentRole of await Meteor.roles
+        .find({ 'children._id': roleName })
+        .fetchAsync()) {
         parentRoles.add(parentRole._id)
-      })
-    })
+      }
+    }
 
     parentRoles.delete(role._id)
 
@@ -542,21 +688,27 @@ Object.assign(Roles, {
    *
    * @method _getInheritedRoleNames
    * @param {object} role The role object
+   * @returns {Promise}
    * @private
    * @static
    */
-  _getInheritedRoleNames: function (role) {
+  _getInheritedRoleNamesAsync: async function (role) {
     const inheritedRoles = new Set()
     const nestedRoles = new Set([role])
 
-    nestedRoles.forEach(r => {
-      const roles = Meteor.roles.find({ _id: { $in: r.children.map(r => r._id) } }, { fields: { children: 1 } }).fetch()
+    for (const r of nestedRoles) {
+      const roles = await Meteor.roles
+        .find(
+          { _id: { $in: r.children.map((r) => r._id) } },
+          { fields: { children: 1 } }
+        )
+        .fetchAsync()
 
-      roles.forEach(r2 => {
+      for (const r2 of roles) {
         inheritedRoles.add(r2._id)
         nestedRoles.add(r2)
-      })
-    })
+      }
+    }
 
     return [...inheritedRoles]
   },
@@ -565,23 +717,24 @@ Object.assign(Roles, {
    * Remove users from assigned roles.
    *
    * @example
-   *     Roles.removeUsersFromRoles(userId, 'admin')
-   *     Roles.removeUsersFromRoles([userId, user2], ['editor'])
-   *     Roles.removeUsersFromRoles(userId, ['user'], 'group1')
+   *     await Roles.removeUsersFromRolesAsync(userId, 'admin')
+   *     await Roles.removeUsersFromRolesAsync([userId, user2], ['editor'])
+   *     await Roles.removeUsersFromRolesAsync(userId, ['user'], 'group1')
    *
-   * @method removeUsersFromRoles
+   * @method removeUsersFromRolesAsync
    * @param {Array|String} users User ID(s) or object(s) with an `_id` field.
    * @param {Array|String} roles Name(s) of roles to remove users from. Roles have to exist.
    * @param {Object|String} [options] Options:
    *   - `scope`: name of the scope, or `null` for the global role
    *   - `anyScope`: if set, role can be in any scope (`scope` option is ignored)
+   * @returns {Promise}
    *
    * Alternatively, it can be a scope name string.
    * @static
    */
-  removeUsersFromRoles: function (users, roles, options) {
-    if (!users) throw new Error('Missing \'users\' param.')
-    if (!roles) throw new Error('Missing \'roles\' param.')
+  removeUsersFromRolesAsync: async function (users, roles, options) {
+    if (!users) throw new Error("Missing 'users' param.")
+    if (!roles) throw new Error("Missing 'roles' param.")
 
     options = Roles._normalizeOptions(options)
 
@@ -591,10 +744,10 @@ Object.assign(Roles, {
 
     Roles._checkScopeName(options.scope)
 
-    users.forEach(function (user) {
+    for (const user of users) {
       if (!user) return
 
-      roles.forEach(function (role) {
+      for (const role of roles) {
         let id
         if (typeof user === 'object') {
           id = user._id
@@ -602,24 +755,25 @@ Object.assign(Roles, {
           id = user
         }
 
-        Roles._removeUserFromRole(id, role, options)
-      })
-    })
+        await Roles._removeUserFromRoleAsync(id, role, options)
+      }
+    }
   },
 
   /**
    * Remove one user from one role.
    *
-   * @method _removeUserFromRole
+   * @method _removeUserFromRoleAsync
    * @param {String} userId The user ID.
    * @param {String} roleName Name of the role to add the user to. The role have to exist.
    * @param {Object} options Options:
    *   - `scope`: name of the scope, or `null` for the global role
    *   - `anyScope`: if set, role can be in any scope (`scope` option is ignored)
+   * @returns {Promise}
    * @private
    * @static
    */
-  _removeUserFromRole: function (userId, roleName, options) {
+  _removeUserFromRoleAsync: async function (userId, roleName, options) {
     Roles._checkRoleName(roleName)
     Roles._checkScopeName(options.scope)
 
@@ -634,7 +788,7 @@ Object.assign(Roles, {
       selector.scope = options.scope
     }
 
-    Meteor.roleAssignment.remove(selector)
+    await Meteor.roleAssignment.removeAsync(selector)
   },
 
   /**
@@ -642,17 +796,17 @@ Object.assign(Roles, {
    *
    * @example
    *     // global roles
-   *     Roles.userIsInRole(user, 'admin')
-   *     Roles.userIsInRole(user, ['admin','editor'])
-   *     Roles.userIsInRole(userId, 'admin')
-   *     Roles.userIsInRole(userId, ['admin','editor'])
+   *     await Roles.userIsInRoleAsync(user, 'admin')
+   *     await Roles.userIsInRoleAsync(user, ['admin','editor'])
+   *     await Roles.userIsInRoleAsync(userId, 'admin')
+   *     await Roles.userIsInRoleAsync(userId, ['admin','editor'])
    *
    *     // scope roles (global roles are still checked)
-   *     Roles.userIsInRole(user, 'admin', 'group1')
-   *     Roles.userIsInRole(userId, ['admin','editor'], 'group1')
-   *     Roles.userIsInRole(userId, ['admin','editor'], {scope: 'group1'})
+   *     await Roles.userIsInRoleAsync(user, 'admin', 'group1')
+   *     await Roles.userIsInRoleAsync(userId, ['admin','editor'], 'group1')
+   *     await Roles.userIsInRoleAsync(userId, ['admin','editor'], {scope: 'group1'})
    *
-   * @method userIsInRole
+   * @method userIsInRoleAsync
    * @param {String|Object} user User ID or an actual user object.
    * @param {Array|String} roles Name of role or an array of roles to check against. If array,
    *                             will return `true` if user is in _any_ role.
@@ -663,27 +817,29 @@ Object.assign(Roles, {
    *   - `anyScope`: if set, role can be in any scope (`scope` option is ignored)
    *
    * Alternatively, it can be a scope name string.
-   * @return {Boolean} `true` if user is in _any_ of the target roles
+   * @return {Promise<Boolean>} `true` if user is in _any_ of the target roles
    * @static
    */
-  userIsInRole: function (user, roles, options) {
-    var id
-    var selector
+  userIsInRoleAsync: async function (user, roles, options) {
+    let id
 
     options = Roles._normalizeOptions(options)
 
     // ensure array to simplify code
     if (!Array.isArray(roles)) roles = [roles]
 
-    roles = roles.filter(r => r != null)
+    roles = roles.filter((r) => r != null)
 
     if (!roles.length) return false
 
     Roles._checkScopeName(options.scope)
 
-    options = Object.assign({
-      anyScope: false
-    }, options)
+    options = Object.assign(
+      {
+        anyScope: false
+      },
+      options
+    )
 
     if (user && typeof user === 'object') {
       id = user._id
@@ -694,7 +850,7 @@ Object.assign(Roles, {
     if (!id) return false
     if (typeof id !== 'string') return false
 
-    selector = {
+    const selector = {
       'user._id': id
     }
 
@@ -702,17 +858,22 @@ Object.assign(Roles, {
       selector.scope = { $in: [options.scope, null] }
     }
 
-    return roles.some((roleName) => {
+    const res = await asyncSome(roles, async (roleName) => {
       selector['inheritedRoles._id'] = roleName
-
-      return Meteor.roleAssignment.find(selector, { limit: 1 }).count() > 0
+      const out =
+        (await Meteor.roleAssignment
+          .find(selector, { limit: 1 })
+          .countAsync()) > 0
+      return out
     })
+
+    return res
   },
 
   /**
    * Retrieve user's roles.
    *
-   * @method getRolesForUser
+   * @method getRolesForUserAsync
    * @param {String|Object} user User ID or an actual user object.
    * @param {Object|String} [options] Options:
    *   - `scope`: name of scope to provide roles for; if not specified, global roles are returned
@@ -721,17 +882,14 @@ Object.assign(Roles, {
    *   - `onlyAssigned`: return only assigned roles and not automatically inferred (like subroles)
    *   - `fullObjects`: return full roles objects (`true`) or just names (`false`) (`onlyAssigned` option is ignored) (default `false`)
    *     If you have a use-case for this option, please file a feature-request. You shouldn't need to use it as it's
-   *     result strongly dependant on the internal data structure of this plugin.
+   *     result strongly dependent on the internal data structure of this plugin.
    *
    * Alternatively, it can be a scope name string.
-   * @return {Array} Array of user's roles, unsorted.
+   * @return {Promise<Array>} Array of user's roles, unsorted.
    * @static
    */
-  getRolesForUser: function (user, options) {
-    var id
-    var selector
-    var filter
-    var roles
+  getRolesForUserAsync: async function (user, options) {
+    let id
 
     options = Roles._normalizeOptions(options)
 
@@ -752,11 +910,11 @@ Object.assign(Roles, {
 
     if (!id) return []
 
-    selector = {
+    const selector = {
       'user._id': id
     }
 
-    filter = {
+    const filter = {
       fields: { 'inheritedRoles._id': 1 }
     }
 
@@ -777,27 +935,31 @@ Object.assign(Roles, {
       delete filter.fields
     }
 
-    roles = Meteor.roleAssignment.find(selector, filter).fetch()
+    const roles = await Meteor.roleAssignment.find(selector, filter).fetchAsync()
 
     if (options.fullObjects) {
       return roles
     }
 
-    return [...new Set(roles.reduce((rev, current) => {
-      if (current.inheritedRoles) {
-        return rev.concat(current.inheritedRoles.map(r => r._id))
-      } else if (current.role) {
-        rev.push(current.role._id)
-      }
-      return rev
-    }, []))]
+    return [
+      ...new Set(
+        roles.reduce((rev, current) => {
+          if (current.inheritedRoles) {
+            return rev.concat(current.inheritedRoles.map((r) => r._id))
+          } else if (current.role) {
+            rev.push(current.role._id)
+          }
+          return rev
+        }, [])
+      )
+    ]
   },
 
   /**
    * Retrieve cursor of all existing roles.
    *
    * @method getAllRoles
-   * @param {Object} queryOptions Options which are passed directly
+   * @param {Object} [queryOptions] Options which are passed directly
    *                                through to `Meteor.roles.find(query, options)`.
    * @return {Cursor} Cursor of existing roles.
    * @static
@@ -813,7 +975,7 @@ Object.assign(Roles, {
    *
    * Options:
    *
-   * @method getUsersInRole
+   * @method getUsersInRoleAsync
    * @param {Array|String} roles Name of role or an array of roles. If array, users
    *                             returned will have at least one of the roles
    *                             specified but need not have _all_ roles.
@@ -829,15 +991,18 @@ Object.assign(Roles, {
    * Alternatively, it can be a scope name string.
    * @param {Object} [queryOptions] Options which are passed directly
    *                                through to `Meteor.users.find(query, options)`
-   * @return {Cursor} Cursor of users in roles.
+   * @return {Promise<Cursor>} Cursor of users in roles.
    * @static
    */
-  getUsersInRole: function (roles, options, queryOptions) {
-    var ids
+  getUsersInRoleAsync: async function (roles, options, queryOptions) {
+    const ids = (
+      await Roles.getUserAssignmentsForRole(roles, options).fetchAsync()
+    ).map((a) => a.user._id)
 
-    ids = Roles.getUserAssignmentsForRole(roles, options).fetch().map(a => a.user._id)
-
-    return Meteor.users.find({ _id: { $in: ids } }, ((options && options.queryOptions) || queryOptions) || {})
+    return Meteor.users.find(
+      { _id: { $in: ids } },
+      (options && options.queryOptions) || queryOptions || {}
+    )
   },
 
   /**
@@ -856,7 +1021,7 @@ Object.assign(Roles, {
    *   - `anyScope`: if set, role can be in any scope (`scope` option is ignored)
    *   - `queryOptions`: options which are passed directly
    *     through to `Meteor.roleAssignment.find(query, options)`
-   *
+
    * Alternatively, it can be a scope name string.
    * @return {Cursor} Cursor of user assignments for roles.
    * @static
@@ -864,10 +1029,13 @@ Object.assign(Roles, {
   getUserAssignmentsForRole: function (roles, options) {
     options = Roles._normalizeOptions(options)
 
-    options = Object.assign({
-      anyScope: false,
-      queryOptions: {}
-    }, options)
+    options = Object.assign(
+      {
+        anyScope: false,
+        queryOptions: {}
+      },
+      options
+    )
 
     return Roles._getUsersInRoleCursor(roles, options, options.queryOptions)
   },
@@ -891,25 +1059,29 @@ Object.assign(Roles, {
    * @static
    */
   _getUsersInRoleCursor: function (roles, options, filter) {
-    var selector
-
     options = Roles._normalizeOptions(options)
 
-    options = Object.assign({
-      anyScope: false,
-      onlyScoped: false
-    }, options)
+    options = Object.assign(
+      {
+        anyScope: false,
+        onlyScoped: false
+      },
+      options
+    )
 
     // ensure array to simplify code
     if (!Array.isArray(roles)) roles = [roles]
 
     Roles._checkScopeName(options.scope)
 
-    filter = Object.assign({
-      fields: { 'user._id': 1 }
-    }, filter)
+    filter = Object.assign(
+      {
+        fields: { 'user._id': 1 }
+      },
+      filter
+    )
 
-    selector = {
+    const selector = {
       'inheritedRoles._id': { $in: roles }
     }
 
@@ -927,32 +1099,35 @@ Object.assign(Roles, {
   /**
    * Deprecated. Use `getScopesForUser` instead.
    *
-   * @method getGroupsForUser
+   * @method getGroupsForUserAsync
+   * @returns {Promise<Array>}
    * @static
    * @deprecated
    */
-  getGroupsForUser: function (...args) {
+  getGroupsForUserAsync: async function (...args) {
     if (!getGroupsForUserDeprecationWarning) {
       getGroupsForUserDeprecationWarning = true
-      console && console.warn('getGroupsForUser has been deprecated. Use getScopesForUser instead.')
+      console &&
+        console.warn(
+          'getGroupsForUser has been deprecated. Use getScopesForUser instead.'
+        )
     }
 
-    return Roles.getScopesForUser(...args)
+    return await Roles.getScopesForUser(...args)
   },
 
   /**
    * Retrieve users scopes, if any.
    *
-   * @method getScopesForUser
+   * @method getScopesForUserAsync
    * @param {String|Object} user User ID or an actual user object.
    * @param {Array|String} [roles] Name of roles to restrict scopes to.
    *
-   * @return {Array} Array of user's scopes, unsorted.
+   * @return {Promise<Array>} Array of user's scopes, unsorted.
    * @static
    */
-  getScopesForUser: function (user, roles) {
-    var scopes
-    var id
+  getScopesForUserAsync: async function (user, roles) {
+    let id
 
     if (roles && !Array.isArray(roles)) roles = [roles]
 
@@ -973,7 +1148,11 @@ Object.assign(Roles, {
       selector['inheritedRoles._id'] = { $in: roles }
     }
 
-    scopes = Meteor.roleAssignment.find(selector, { fields: { scope: 1 } }).fetch().map(obi => obi.scope)
+    const scopes = (
+      await Meteor.roleAssignment
+        .find(selector, { fields: { scope: 1 } })
+        .fetchAsync()
+    ).map((obi) => obi.scope)
 
     return [...new Set(scopes)]
   },
@@ -983,13 +1162,14 @@ Object.assign(Roles, {
    *
    * Roles assigned with a given scope are changed to be under the new scope.
    *
-   * @method renameScope
+   * @method renameScopeAsync
    * @param {String} oldName Old name of a scope.
    * @param {String} newName New name of a scope.
+   * @returns {Promise}
    * @static
    */
-  renameScope: function (oldName, newName) {
-    var count
+  renameScopeAsync: async function (oldName, newName) {
+    let count
 
     Roles._checkScopeName(oldName)
     Roles._checkScopeName(newName)
@@ -997,13 +1177,17 @@ Object.assign(Roles, {
     if (oldName === newName) return
 
     do {
-      count = Meteor.roleAssignment.update({
-        scope: oldName
-      }, {
-        $set: {
-          scope: newName
-        }
-      }, { multi: true })
+      count = await Meteor.roleAssignment.updateAsync(
+        {
+          scope: oldName
+        },
+        {
+          $set: {
+            scope: newName
+          }
+        },
+        { multi: true }
+      )
     } while (count > 0)
   },
 
@@ -1012,14 +1196,15 @@ Object.assign(Roles, {
    *
    * Roles assigned with a given scope are removed.
    *
-   * @method removeScope
+   * @method removeScopeAsync
    * @param {String} name The name of a scope.
+   * @returns {Promise}
    * @static
    */
-  removeScope: function (name) {
+  removeScopeAsync: async function (name) {
     Roles._checkScopeName(name)
 
-    Meteor.roleAssignment.remove({ scope: name })
+    await Meteor.roleAssignment.removeAsync({ scope: name })
   },
 
   /**
@@ -1031,8 +1216,12 @@ Object.assign(Roles, {
    * @static
    */
   _checkRoleName: function (roleName) {
-    if (!roleName || typeof roleName !== 'string' || roleName.trim() !== roleName) {
-      throw new Error('Invalid role name \'' + roleName + '\'.')
+    if (
+      !roleName ||
+      typeof roleName !== 'string' ||
+      roleName.trim() !== roleName
+    ) {
+      throw new Error("Invalid role name '" + roleName + "'.")
     }
   },
 
@@ -1041,12 +1230,13 @@ Object.assign(Roles, {
    *
    * WARNING: If you check this on the client, please make sure all roles are published.
    *
-   * @method isParentOf
+   * @method isParentOfAsync
    * @param {String} parentRoleName The role you want to research.
    * @param {String} childRoleName The role you expect to be among the children of parentRoleName.
+   * @returns {Promise}
    * @static
    */
-  isParentOf: function (parentRoleName, childRoleName) {
+  isParentOfAsync: async function (parentRoleName, childRoleName) {
     if (parentRoleName === childRoleName) {
       return true
     }
@@ -1058,20 +1248,20 @@ Object.assign(Roles, {
     Roles._checkRoleName(parentRoleName)
     Roles._checkRoleName(childRoleName)
 
-    var rolesToCheck = [parentRoleName]
+    let rolesToCheck = [parentRoleName]
     while (rolesToCheck.length !== 0) {
-      var roleName = rolesToCheck.pop()
+      const roleName = rolesToCheck.pop()
 
       if (roleName === childRoleName) {
         return true
       }
 
-      var role = Meteor.roles.findOne({ _id: roleName })
+      const role = await Meteor.roles.findOneAsync({ _id: roleName })
 
       // This should not happen, but this is a problem to address at some other time.
       if (!role) continue
 
-      rolesToCheck = rolesToCheck.concat(role.children.map(r => r._id))
+      rolesToCheck = rolesToCheck.concat(role.children.map((r) => r._id))
     }
 
     return false
@@ -1127,8 +1317,12 @@ Object.assign(Roles, {
   _checkScopeName: function (scopeName) {
     if (scopeName === null) return
 
-    if (!scopeName || typeof scopeName !== 'string' || scopeName.trim() !== scopeName) {
-      throw new Error('Invalid scope name \'' + scopeName + '\'.')
+    if (
+      !scopeName ||
+      typeof scopeName !== 'string' ||
+      scopeName.trim() !== scopeName
+    ) {
+      throw new Error("Invalid scope name '" + scopeName + "'.")
     }
   }
 })